Split the remaining classes into separate files
